+++ src/client/app/app-routing.module.ts
@@ -8,7 +8,7 @@ import { UserSignupComponent } from './admin/user-management/user-signup.compone
 import { UpdateUserComponent } from './admin/user-management/edit-user.component';
 import { InvalidResourceComponent } from './core/invalid-resource.component';
 import { InvalidCertificateComponent } from './admin/authentication/invalid-certificate.component';
-import { AuditComponent } from './audit/components/audit.client.component';
+import { AuditComponent } from './audit/audit.component';
 import { UserEuaComponent } from './admin/end-user-agreement/user-eua.component';
 import { NoAccessComponent } from './admin/authentication/no-access.component';
 import { LoggedInComponent } from './core/logged-in.component';
