+++ src/client/app/audit/audit.module.ts
@@ -5,13 +5,13 @@ import { NgModule } from '@angular/core';
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 import { BootstrapModalModule, Modal } from 'angular2-modal/plugins/bootstrap';
 
-import { AuditViewChangeModal, AuditViewDetailModal } from './components/audit-view-change.client.component';
-import { AuditService } from './services/audit.client.service';
-import { AuditComponent } from './components/audit.client.component';
+import { AuditViewChangeModal, AuditViewDetailModal } from './audit-view-change.component';
+import { AuditService } from './audit.service';
+import { AuditComponent } from './audit.component';
 import {
 	AuditObjectComponent, UrlAudit, UserAuthenticationAudit, UserAudit,
 	auditObjects
-} from './components/audit-object.client.component';
+} from './audit-object.component';
 import { UtilModule } from '../shared/util.module';
 import { UserService } from '../admin/users.service';
