+++ src/client/app/audit/audit.service.ts
@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Response} from '@angular/http';
 import {Observable} from 'rxjs/Observable';
-import { AsyHttp, HttpOptions } from '../../shared/asy-http.service';
-import { PagingOptions } from '../../shared/pager.component';
+import { AsyHttp, HttpOptions } from '../shared/asy-http.service';
+import { PagingOptions } from '../shared/pager.component';
 
 @Injectable()
 export class AuditService {
