+++ src/client/app/audit/templates/group.audit.view.html
