+++ src/client/app/audit/templates/user-authentication.audit.view.html
