+++ src/client/app/audit/templates/user.audit.view.html
